floating damage text

+++ Assets/WeaponPerks/WeaponData.cs
@@ -21,11 +21,11 @@ public class WeaponData : ScriptableObject
 {
     
     public string weaponName;
-    public float baseDamage;
+    public int baseDamage;
     public float baseReloadTime;
     public List<WeaponPerk> possiblePerks; // List of possible perks for this weapon
-    public List<float> perkValues; // Additional values for each perk
-    public List<float> perkDurations;
+    public List<int> perkValues; // Additional values for each perk
+    public List<int> perkDurations;
     public float range;
     public int basereserveAmmo;
     public int basemaxAmmoInMag;
